remove karma move to mocha [closes #14]
