--- karma.conf.js
@@ -1,5 +1,0 @@
-var karmaConfig = require('./gulpfile.js/config/karma')
-
-module.exports = function(config) {
-  config.set(karmaConfig)
-}
