+++ package.json
@@ -10,19 +10,14 @@
     "url" : "https://github.com/artisin/teFlow"
   },
   "scripts": {
-    "test": "gulp test"
+    "test": "mocha --recursive __tests__/"
   },
   "keywords": [
-    "artisin",
     "te",
-    "javascript",
-    "stream",
-    "argument stream",
     "flow",
     "control",
     "wrapper",
     "function",
-    "functional",
     "pipeline"
   ],
   "dependencies":{
@@ -36,22 +31,12 @@
     "gulp": "^3.9.0",
     "gulp-sequence": "^0.4.1",
     "gulp-jade": "^1.1.0",
-    "karma": "^0.13.9",
     "webpack": "^1.11.0",
     "browser-sync": "^2.8.2",
     "babel-loader": "^5.3.2",
     "source-map-support": "^0.2.10",
-    "karma-cli": "^0.1.*",
-    "karma-firefox-launcher": "^0.1.6",
-    "karma-mocha": "^0.2.0",
-    "karma-nyan-reporter": "^0.2.2",
-    "karma-sinon-chai": "^1.0.0",
-    "karma-phantomjs-launcher": "^0.2.1",
-    "karma-sourcemap-loader": "^0.3.5",
-    "karma-webpack": "^1.7.0",
-    "karma-nyan-reporter": "^0.2.2",
-    "phantomjs": "1.9.18",
     "expect.js": "0.3.1",
+    "mocha": "3.2.0",
     "webpack-umd-external": "1.0.2",
     "require-dir": "^0.3.0",
     "colors": "^1.1.2",
